requirements.txt
google-genai
google-generativeai==0.3.1
pydantic==2.11.3
python-dotenv
thefuzz
uvicorn[standard]
pydantic[email]
fastapi